# Add perf sensitive inline attributes in interpreter hot path