--- a/src/v1/engine/code_map.rs
+++ b/src/v1/engine/code_map.rs
@@ -183,7 +183,7 @@ impl<'a> ResolvedFuncBody<'a> {
     /// Visits the corresponding [`Instruction`] method from the given visitor.
     ///
     /// Returns the visitor's outcome value.
-    #[inline]
+    #[inline(always)]
     pub fn visit<T>(&self, index: usize, mut visitor: T) -> T::Outcome
     where
         T: VisitInstruction,