--- a/src/v1/engine/exec_context.rs
+++ b/src/v1/engine/exec_context.rs
@@ -57,6 +57,12 @@ impl<'engine, 'func> ExecutionContext<'engine, 'func> {
         })
     }
 
+    /// Executes the current function frame.
+    ///
+    /// # Note
+    ///
+    /// This executes instructions sequentially until either the function
+    /// calls into another function or the function returns to its caller.
     #[inline(always)]
     pub fn execute_frame(
         &mut self,